// client/src/models/bois/ContrePlaque.ts
export interface PostContrePlaque {
  type: string;
  marque: string;
  fornisseur: string;
  face: string;
  pieces: number;
  long: number;
  larg: number;
  epaisseur: number;
  prix_unity: number;
  date_creation: string;
}
export interface GetContrePlaque {
  id: number;
  type: string;
  marque: string;
  fornisseur: string;
  face: string;
  pieces: number;
  long: number;
  larg: number;
  epaisseur: number;
  prix_unity: number;
  date_creation: string;
}